setAbandonedBasketCookie to utils and set on gw checkout
